+++ src/common/requests/calendar/find-calendar-by-id.request.ts
@@ -1,6 +1,6 @@
 import { Transform } from 'class-transformer';
 import { IsNumber } from 'class-validator';
-import { numberTransform } from '../../transform/number.transform';
+import { numberTransform } from '../transform/number.transform';
 
 export class FindCalendarByIdRequest {
   @IsNumber()
